Separate Apache Http Client request processing from client and handler
Introduce a new API to process Apache Http Client's HttpMethod using the
existing configurations. Its primary responsibility is to process the
method with configurations and execute the method. Handler will handle how
to process the request. When adapting to other APIs the ThreadLocal can be
used to pass through the ClientRequest.

In order to use the API we needed to expose componentProviderFactory to its
children so that they can use it to inject into other components manually
wherever required. Now its required to inject beans into the processor.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>