Adding support for generating XML Schema from JAX-B elements, this isn't enabled by default currently you have to define a WadlGeneratorConfig that makes use of the main class:
public static class SchemaWadlGeneratorConfig extends WadlGeneratorConfig {
    @Override
    public List<WadlGeneratorDescription> configure() {
        return generator(WadlGeneratorJAXBGrammarGenerator.class).descriptions();
    }
}

This class will look at the resources and generate a schema for any element that is a @XmlRootElement that is either a direct reference in a return value or entity parameter, a generic parameter of a JResponse return value or referenced indirectly by @XmlSeeAlso on a resource class.

When installed the generated schemas are laid out as follows, and they are referenced from both the root application.wadl and from the WADL generated on a OPTIONS request.

%root%/application.wadl/xsd0.xsd
%root%/application.wadl/xsd1.xsd etc

WadlGenerator now has an extra method to provide the external metadata, the lifecycle has also changed so a new instance is created on each generation so state can be passed between the different methods.

TODO Items:

Caching in WadlApplicationContextImpl
Support for indirect references via List<>
Make the generation enabled by default.

git-svn-id: https://svn.java.net/svn/jersey~svn/trunk/jersey@5180 2fda5173-fe30-0410-90bd-98ff5453417f